chore: (apps/web): Reorganize files
